Remove ADC_ASSERT, ADC_MALLOC and ADC_FREE macros
I'm dumb and these can't be overridden if the c implementation
is included in a program build. That approach only works with single
file header libraries. I like the single c/h file approach for now,
and almost all of my projects don't need to override these things, so
I've removed the macros. I've left the max error definition, as that
could just be overidden when compiling using a cli flag.